@@ -19,44 +19,44 @@ import * as i14 from "./components/onboarding/onboarding-detail/onboarding-detai
 import * as i15 from "./components/onboarding/loading/loading.component";
 import * as i16 from "./components/loading/loading.component";
 import * as i17 from "./components/dialog-confirmation/dialog-confirmation.component";
-import * as i18 from "./pages/template-profile-form/template-profile-form.component";
-import * as i19 from "@angular/forms";
-import * as i20 from "@angular/material/form-field";
-import * as i21 from "@angular/material/button";
-import * as i22 from "@angular/common";
-import * as i23 from "@angular/platform-browser";
-import * as i24 from "@angular/platform-browser/animations";
-import * as i25 from "@angular/material/checkbox";
-import * as i26 from "@angular/material/icon";
-import * as i27 from "@angular/material/select";
-import * as i28 from "@angular/material/input";
-import * as i29 from "@angular/material/chips";
-import * as i30 from "@angular/material/grid-list";
-import * as i31 from "@angular/material/radio";
-import * as i32 from "@angular/material/paginator";
-import * as i33 from "@angular/cdk/table";
-import * as i34 from "@angular/material/sort";
-import * as i35 from "@angular/material/table";
-import * as i36 from "@angular/router";
-import * as i37 from "@nicky-lenaers/ngx-scroll-to";
-import * as i38 from "@angular/cdk/drag-drop";
-import * as i39 from "@angular/material/card";
-import * as i40 from "@angular/material/dialog";
-import * as i41 from "@angular/material/stepper";
-import * as i42 from "@angular/material/progress-bar";
-import * as i43 from "@angular/material/tooltip";
-import * as i44 from "table";
-import * as i45 from "@angular/material/tabs";
-import * as i46 from "ngx-intl-tel-input";
-import * as i47 from "@angular/material/slide-toggle";
-import * as i48 from "@angular/material/datepicker";
-import * as i49 from "@angular/material/core";
-import * as i50 from "@angular/material/bottom-sheet";
-import * as i51 from "@angular/common/http";
+import * as i18 from "@angular/forms";
+import * as i19 from "@angular/material/form-field";
+import * as i20 from "@angular/material/button";
+import * as i21 from "@angular/common";
+import * as i22 from "@angular/platform-browser";
+import * as i23 from "@angular/platform-browser/animations";
+import * as i24 from "@angular/material/checkbox";
+import * as i25 from "@angular/material/icon";
+import * as i26 from "@angular/material/select";
+import * as i27 from "@angular/material/input";
+import * as i28 from "@angular/material/chips";
+import * as i29 from "@angular/material/grid-list";
+import * as i30 from "@angular/material/radio";
+import * as i31 from "@angular/material/paginator";
+import * as i32 from "@angular/cdk/table";
+import * as i33 from "@angular/material/sort";
+import * as i34 from "@angular/material/table";
+import * as i35 from "@angular/router";
+import * as i36 from "@nicky-lenaers/ngx-scroll-to";
+import * as i37 from "@angular/cdk/drag-drop";
+import * as i38 from "@angular/material/card";
+import * as i39 from "@angular/material/dialog";
+import * as i40 from "@angular/material/stepper";
+import * as i41 from "@angular/material/progress-bar";
+import * as i42 from "@angular/material/tooltip";
+import * as i43 from "ngx-table";
+import * as i44 from "@angular/material/tabs";
+import * as i45 from "ngx-intl-tel-input";
+import * as i46 from "@angular/material/slide-toggle";
+import * as i47 from "@angular/material/datepicker";
+import * as i48 from "@angular/material/core";
+import * as i49 from "@angular/material/bottom-sheet";
+import * as i50 from "@angular/common/http";
 export declare const onBoardingRoutes: Routes;
 export declare class OnboardingModule {
     constructor(parentModule?: OnboardingModule);
     static forRoot(config: OnboardingInterface): ModuleWithProviders<OnboardingModule>;
-    static ɵmod: i0.ɵɵNgModuleDefWithMeta<OnboardingModule, [typeof i1.MaterialElevationDirective, typeof i2.OnboardingComponent, typeof i3.TemplateProfileCreateComponent, typeof i4.OnboardingCreateComponent, typeof i5.SubOnboardingCreateComponent, typeof i6.TemplateProfileListComponent, typeof i7.SubTemplateProfileComponent, typeof i8.ExitConfirmComponent, typeof i9.OnboardingListComponent, typeof i10.TypeComponent, typeof i11.OnboardingTableComponent, typeof i12.DialogCreateTemplateComponent, typeof i13.DialogCreateOnboardingComponent, typeof i14.OnboardingDetailComponent, typeof i15.LoadingComponent, typeof i16.LoadingPageComponent, typeof i17.DialogConfirmation, typeof i14.BottomSheet, typeof i18.TemplateProfileFormComponent], [typeof i19.FormsModule, typeof i20.MatFormFieldModule, typeof i21.MatButtonModule, typeof i22.CommonModule, typeof i23.BrowserModule, typeof i24.BrowserAnimationsModule, typeof i19.FormsModule, typeof i19.ReactiveFormsModule, typeof i25.MatCheckboxModule, typeof i26.MatIconModule, typeof i27.MatSelectModule, typeof i28.MatInputModule, typeof i29.MatChipsModule, typeof i30.MatGridListModule, typeof i31.MatRadioModule, typeof i32.MatPaginatorModule, typeof i33.CdkTableModule, typeof i34.MatSortModule, typeof i35.MatTableModule, typeof i36.RouterModule, typeof i37.ScrollToModule, typeof i38.DragDropModule, typeof i39.MatCardModule, typeof i40.MatDialogModule, typeof i41.MatStepperModule, typeof i42.MatProgressBarModule, typeof i43.MatTooltipModule, typeof i44.TableModule, typeof i45.MatTabsModule, typeof i46.NgxIntlTelInputModule, typeof i47.MatSlideToggleModule, typeof i48.MatDatepickerModule, typeof i49.MatNativeDateModule, typeof i50.MatBottomSheetModule], [typeof i2.OnboardingComponent, typeof i51.HttpClientModule, typeof i3.TemplateProfileCreateComponent, typeof i6.TemplateProfileListComponent, typeof i9.OnboardingListComponent, typeof i4.OnboardingCreateComponent, typeof i13.DialogCreateOnboardingComponent, typeof i36.RouterModule, typeof i45.MatTabsModule]>;
-    static ɵinj: i0.ɵɵInjectorDef<OnboardingModule>;
+    static ɵfac: i0.ɵɵFactoryDeclaration<OnboardingModule, [{ optional: true; skipSelf: true; }]>;
+    static ɵmod: i0.ɵɵNgModuleDeclaration<OnboardingModule, [typeof i1.MaterialElevationDirective, typeof i2.OnboardingComponent, typeof i3.TemplateProfileCreateComponent, typeof i4.OnboardingCreateComponent, typeof i5.SubOnboardingCreateComponent, typeof i6.TemplateProfileListComponent, typeof i7.SubTemplateProfileComponent, typeof i8.ExitConfirmComponent, typeof i9.OnboardingListComponent, typeof i10.TypeComponent, typeof i11.OnboardingTableComponent, typeof i12.DialogCreateTemplateComponent, typeof i13.DialogCreateOnboardingComponent, typeof i14.OnboardingDetailComponent, typeof i15.LoadingComponent, typeof i16.LoadingPageComponent, typeof i17.DialogConfirmation, typeof i14.BottomSheet], [typeof i18.FormsModule, typeof i19.MatFormFieldModule, typeof i20.MatButtonModule, typeof i21.CommonModule, typeof i22.BrowserModule, typeof i23.BrowserAnimationsModule, typeof i18.FormsModule, typeof i18.ReactiveFormsModule, typeof i24.MatCheckboxModule, typeof i25.MatIconModule, typeof i26.MatSelectModule, typeof i27.MatInputModule, typeof i28.MatChipsModule, typeof i29.MatGridListModule, typeof i30.MatRadioModule, typeof i31.MatPaginatorModule, typeof i32.CdkTableModule, typeof i33.MatSortModule, typeof i34.MatTableModule, typeof i35.RouterModule, typeof i36.ScrollToModule, typeof i37.DragDropModule, typeof i38.MatCardModule, typeof i39.MatDialogModule, typeof i40.MatStepperModule, typeof i41.MatProgressBarModule, typeof i42.MatTooltipModule, typeof i43.TableModule, typeof i44.MatTabsModule, typeof i45.NgxIntlTelInputModule, typeof i46.MatSlideToggleModule, typeof i47.MatDatepickerModule, typeof i48.MatNativeDateModule, typeof i49.MatBottomSheetModule, typeof i36.ScrollToModule], [typeof i2.OnboardingComponent, typeof i50.HttpClientModule, typeof i3.TemplateProfileCreateComponent, typeof i6.TemplateProfileListComponent, typeof i9.OnboardingListComponent, typeof i4.OnboardingCreateComponent, typeof i13.DialogCreateOnboardingComponent, typeof i35.RouterModule, typeof i44.MatTabsModule, typeof i36.ScrollToModule]>;
+    static ɵinj: i0.ɵɵInjectorDeclaration<OnboardingModule>;
 }
