@@ -6,6 +6,6 @@ export interface CanComponentDeactivate {
 }
 export declare class CanDeactivateGuard implements CanDeactivate<CanComponentDeactivate> {
     canDeactivate(component: CanComponentDeactivate, route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | Promise<boolean> | Observable<boolean>;
-    static ɵfac: i0.ɵɵFactoryDef<CanDeactivateGuard>;
-    static ɵprov: i0.ɵɵInjectableDef<CanDeactivateGuard>;
+    static ɵfac: i0.ɵɵFactoryDeclaration<CanDeactivateGuard, never>;
+    static ɵprov: i0.ɵɵInjectableDeclaration<CanDeactivateGuard>;
 }
