@@ -3,6 +3,6 @@ import * as i0 from "@angular/core";
 export declare class LoadingPageComponent implements OnInit {
     constructor();
     ngOnInit(): void;
-    static ɵfac: i0.ɵɵFactoryDef<LoadingPageComponent>;
-    static ɵcmp: i0.ɵɵComponentDefWithMeta<LoadingPageComponent, "lib-loading-page", never, {}, {}, never>;
+    static ɵfac: i0.ɵɵFactoryDeclaration<LoadingPageComponent, never>;
+    static ɵcmp: i0.ɵɵComponentDeclaration<LoadingPageComponent, "lib-loading-page", never, {}, {}, never, never>;
 }
