@@ -3,6 +3,6 @@ import * as i0 from "@angular/core";
 export declare class LoadingComponent implements OnInit {
     constructor();
     ngOnInit(): void;
-    static ɵfac: i0.ɵɵFactoryDef<LoadingComponent>;
-    static ɵcmp: i0.ɵɵComponentDefWithMeta<LoadingComponent, "lib-loading", never, {}, {}, never>;
+    static ɵfac: i0.ɵɵFactoryDeclaration<LoadingComponent, never>;
+    static ɵcmp: i0.ɵɵComponentDeclaration<LoadingComponent, "lib-loading", never, {}, {}, never, never>;
 }
