@@ -8,6 +8,6 @@ export declare class OnboardingTableComponent implements OnInit {
     inputSearch: any;
     constructor();
     ngOnInit(): void;
-    static ɵfac: i0.ɵɵFactoryDef<OnboardingTableComponent>;
-    static ɵcmp: i0.ɵɵComponentDefWithMeta<OnboardingTableComponent, "lib-onboarding-table", never, { "_loading_table": "_loading_table"; "data": "data"; "columnDefinitions": "columnDefinitions"; "lang": "lang"; "inputSearch": "inputSearch"; }, {}, never>;
+    static ɵfac: i0.ɵɵFactoryDeclaration<OnboardingTableComponent, never>;
+    static ɵcmp: i0.ɵɵComponentDeclaration<OnboardingTableComponent, "lib-onboarding-table", never, { "_loading_table": "_loading_table"; "data": "data"; "columnDefinitions": "columnDefinitions"; "lang": "lang"; "inputSearch": "inputSearch"; }, {}, never, never>;
 }
